requirements.txt
alembic==1.11.1
beautifulsoup4==4.12.0
discord.py==2.2.2
langcodes==3.3.0
language_data==1.1
pymediawiki==0.7.2
PyNaCl==1.5.0
pytz==2022.7.1
SQLAlchemy==2.0.13
tldextract==3.4.4
validators==0.20.0
